requirements.txt
python-telegram-bot==20.8
python-dotenv
beautifulsoup4==4.13.4
requests